Separated transit_fitting module into gui, io and core components.
