+++ stellarphot/transit_fitting/__init__.py
@@ -1,2 +1,1 @@
-from .exotic import *
-from .model_fit import *
+from .core import *
