+++ stellarphot/transit_fitting/gui/data/exotic-to-mod-known.json
