+++ stellarphot/transit_fitting/gui/data/tic-template-for-exotic-known.json
