+++ stellarphot/transit_fitting/tests/test_transit_model_fit.py
@@ -5,7 +5,7 @@ import numpy as np
 
 batman = pytest.importorskip('batman')
 
-from stellarphot.transit_fitting.model_fit import TransitModelFit
+from stellarphot.transit_fitting import TransitModelFit
 
 
 DEFAULT_TESTING_PARAMS = dict(
